# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Period Poverty API","version":"v1"},"paths":{"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/health":{"get":{"summary":"Health","operationId":"health_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/basket-items/":{"get":{"tags":["Basket Items"],"summary":"List Items","operationId":"list_items_v1_basket_items__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BasketItemRead"},"type":"array","title":"Response List Items V1 Basket Items  Get"}}}}}},"post":{"tags":["Basket Items"],"summary":"Create Item","operationId":"create_item_v1_basket_items__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BasketItemCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BasketItemRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/basket-items/{item_id}":{"get":{"tags":["Basket Items"],"summary":"Get Item","operationId":"get_item_v1_basket_items__item_id__get","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BasketItemRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Basket Items"],"summary":"Update Item","operationId":"update_item_v1_basket_items__item_id__put","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BasketItemUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BasketItemRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Basket Items"],"summary":"Delete Item","operationId":"delete_item_v1_basket_items__item_id__delete","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/price-index/":{"get":{"tags":["Dataset: Price Index"],"summary":"CPI personal care monthly series","operationId":"list_price_index_v1_price_index__get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start month (YYYY-MM) or date (YYYY-MM-DD)","title":"From"},"description":"Start month (YYYY-MM) or date (YYYY-MM-DD)"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End month (YYYY-MM) or date (YYYY-MM-DD)","title":"To"},"description":"End month (YYYY-MM) or date (YYYY-MM-DD)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PriceIndexRead"},"title":"Response List Price Index V1 Price Index  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/pip/uk/{year}":{"get":{"tags":["Dataset: PIP (UK)"],"summary":"UK percentiles for a year (daily PPP)","operationId":"pip_uk_year_v1_pip_uk__year__get","parameters":[{"name":"year","in":"path","required":true,"schema":{"type":"integer","maximum":2100,"minimum":1900,"title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IncomePovertyRead"},"title":"Response Pip Uk Year V1 Pip Uk  Year  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/hygiene/uk":{"get":{"tags":["Dataset: Hygiene (UK)"],"summary":"Latest UK hygiene indicator(s)","operationId":"hygiene_uk_v1_hygiene_uk_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/HygieneAccessRead"},"type":"array","title":"Response Hygiene Uk V1 Hygiene Uk Get"}}}}}}},"/v1/analytics/inflation-trend":{"get":{"tags":["Analytics"],"summary":"CPI personal care trend","operationId":"inflation_trend_v1_analytics_inflation_trend_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"YYYY-MM or YYYY-MM-DD","title":"From"},"description":"YYYY-MM or YYYY-MM-DD"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"YYYY-MM or YYYY-MM-DD","title":"To"},"description":"YYYY-MM or YYYY-MM-DD"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InflationTrendResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/analytics/cost-estimate":{"post":{"tags":["Analytics"],"summary":"Estimate monthly/annual basket cost","description":"Computes:\n  - monthly_cost = sum(unit_price * units_per_month)\n  - annual_cost  = monthly_cost * 12\n\nIf `apply_yoy_cpi=True`, the monthly cost is multiplied by (1 + latest_yoy/100)\nusing the latest record in the PriceIndex table.","operationId":"cost_estimate_v1_analytics_cost_estimate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CostEstimateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CostEstimateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/analytics/cost-burden":{"get":{"tags":["Analytics"],"summary":"Annual cost burden vs PIP welfare","operationId":"cost_burden_v1_analytics_cost_burden_get","parameters":[{"name":"year","in":"query","required":true,"schema":{"type":"integer","maximum":2100,"minimum":1900,"title":"Year"}},{"name":"percentile","in":"query","required":true,"schema":{"type":"integer","maximum":100,"minimum":1,"title":"Percentile"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CostBurdenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/analytics/severity-score":{"get":{"tags":["Analytics"],"summary":"Combine hygiene + affordability into a severity score","operationId":"severity_score_v1_analytics_severity_score_get","parameters":[{"name":"year","in":"query","required":false,"schema":{"type":"integer","maximum":2100,"minimum":1900,"description":"Hygiene indicator year; default 2018 for UK","default":2018,"title":"Year"},"description":"Hygiene indicator year; default 2018 for UK"},{"name":"percentile","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Income percentile for affordability comparison","default":20,"title":"Percentile"},"description":"Income percentile for affordability comparison"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeverityScoreResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BasketItemCreate":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"unit_price":{"type":"number","minimum":0.0,"title":"Unit Price"},"units_per_month":{"type":"number","minimum":0.0,"title":"Units Per Month"},"currency":{"type":"string","maxLength":10,"minLength":1,"title":"Currency","default":"GBP"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["name","unit_price","units_per_month"],"title":"BasketItemCreate","description":"POST body for creating a basket item.","example":{"currency":"GBP","name":"pads pack","notes":"supermarket average","unit_price":2.5,"units_per_month":2}},"BasketItemRead":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"unit_price":{"type":"number","minimum":0.0,"title":"Unit Price"},"units_per_month":{"type":"number","minimum":0.0,"title":"Units Per Month"},"currency":{"type":"string","maxLength":10,"minLength":1,"title":"Currency","default":"GBP"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","unit_price","units_per_month","id"],"title":"BasketItemRead","description":"Response model for a basket item (includes id).","example":{"currency":"GBP","id":1,"name":"pads pack","notes":"supermarket average","unit_price":2.5,"units_per_month":2.0}},"BasketItemUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":255,"minLength":1},{"type":"null"}],"title":"Name"},"unit_price":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Unit Price"},"units_per_month":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Units Per Month"},"currency":{"anyOf":[{"type":"string","maxLength":10,"minLength":1},{"type":"null"}],"title":"Currency"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","title":"BasketItemUpdate","description":"PUT body for updating a basket item (partial, all fields optional).","example":{"unit_price":3.0}},"BasketLine":{"properties":{"name":{"type":"string","title":"Name"},"unit_price":{"type":"number","minimum":0.0,"title":"Unit Price"},"units_per_month":{"type":"number","minimum":0.0,"title":"Units Per Month"},"currency":{"type":"string","title":"Currency","default":"GBP"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["name","unit_price","units_per_month"],"title":"BasketLine"},"CostBurdenResponse":{"properties":{"year":{"type":"integer","title":"Year"},"percentile":{"type":"integer","title":"Percentile"},"annual_cost":{"type":"number","title":"Annual Cost"},"avg_welfare_annual_ppp":{"type":"number","title":"Avg Welfare Annual Ppp"},"burden_ratio":{"type":"number","title":"Burden Ratio"}},"type":"object","required":["year","percentile","annual_cost","avg_welfare_annual_ppp","burden_ratio"],"title":"CostBurdenResponse","description":"Annual cost burden (ratio) compared against PIP annual PPP.","example":{"annual_cost":60.0,"avg_welfare_annual_ppp":7500.0,"burden_ratio":0.008,"percentile":20,"year":2018}},"CostEstimateRequest":{"properties":{"lines":{"anyOf":[{"items":{"$ref":"#/components/schemas/BasketLine"},"type":"array"},{"type":"null"}],"title":"Lines"},"apply_yoy_cpi":{"type":"boolean","title":"Apply Yoy Cpi","default":false}},"type":"object","title":"CostEstimateRequest","description":"POST body for cost estimation. If `lines` is omitted/empty, API uses DB BasketItem rows.","example":{"apply_yoy_cpi":true,"lines":[{"currency":"GBP","name":"pads pack","unit_price":2.5,"units_per_month":2}]}},"CostEstimateResponse":{"properties":{"monthly_cost":{"type":"number","title":"Monthly Cost"},"annual_cost":{"type":"number","title":"Annual Cost"},"used_lines":{"items":{"$ref":"#/components/schemas/BasketLine"},"type":"array","title":"Used Lines"}},"type":"object","required":["monthly_cost","annual_cost","used_lines"],"title":"CostEstimateResponse","example":{"annual_cost":60.0,"monthly_cost":5.0,"used_lines":[{"currency":"GBP","name":"pads pack","unit_price":2.5,"units_per_month":2}]}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HygieneAccessRead":{"properties":{"country":{"type":"string","title":"Country"},"year":{"type":"integer","title":"Year"},"indicator":{"type":"string","title":"Indicator"},"value":{"type":"number","title":"Value"}},"type":"object","required":["country","year","indicator","value"],"title":"HygieneAccessRead","description":"UK hygiene proxy row (bathing facility %).","example":{"country":"United Kingdom","indicator":"bathing_facility","value":99.7,"year":2018}},"IncomePovertyRead":{"properties":{"year":{"type":"integer","title":"Year"},"percentile":{"type":"integer","title":"Percentile"},"avg_welfare":{"type":"number","title":"Avg Welfare"},"welfare_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Welfare Type"}},"type":"object","required":["year","percentile","avg_welfare"],"title":"IncomePovertyRead","description":"UK PIP percentile row (daily PPP in `avg_welfare`).","example":{"avg_welfare":18.75,"percentile":20,"welfare_type":"income","year":2018}},"InflationTrendPoint":{"properties":{"date":{"type":"string","title":"Date"},"cpi_index":{"type":"number","title":"Cpi Index"},"pct_change_yoy":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Pct Change Yoy"}},"type":"object","required":["date","cpi_index"],"title":"InflationTrendPoint"},"InflationTrendResponse":{"properties":{"from":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From"},"to":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To"},"points":{"items":{"$ref":"#/components/schemas/InflationTrendPoint"},"type":"array","title":"Points"}},"type":"object","required":["points"],"title":"InflationTrendResponse","example":{"from":"2018-01","points":[{"cpi_index":100.5,"date":"2018-01-01","pct_change_yoy":2.1},{"cpi_index":100.6,"date":"2018-02-01","pct_change_yoy":2.0}],"to":"2020-12"}},"PriceIndexRead":{"properties":{"date":{"type":"string","title":"Date"},"cpi_index":{"type":"number","title":"Cpi Index"},"pct_change_mom":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Pct Change Mom"},"pct_change_yoy":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Pct Change Yoy"}},"type":"object","required":["date","cpi_index"],"title":"PriceIndexRead","description":"CPI personal care row (monthly).","example":{"cpi_index":119.8,"date":"2025-12-01","pct_change_mom":0.2,"pct_change_yoy":3.4}},"SeverityScoreResponse":{"properties":{"year":{"type":"integer","title":"Year"},"annual_cost":{"type":"number","title":"Annual Cost"},"avg_welfare_annual_ppp":{"type":"number","title":"Avg Welfare Annual Ppp"},"burden_ratio":{"type":"number","title":"Burden Ratio"},"hygiene_value_pct":{"type":"number","title":"Hygiene Value Pct"},"hygiene_severity":{"type":"number","title":"Hygiene Severity"},"combined_severity":{"type":"number","title":"Combined Severity"}},"type":"object","required":["year","annual_cost","avg_welfare_annual_ppp","burden_ratio","hygiene_value_pct","hygiene_severity","combined_severity"],"title":"SeverityScoreResponse","description":"Simple composite score combining affordability (burden) and hygiene severity.","example":{"annual_cost":60.0,"avg_welfare_annual_ppp":7500.0,"burden_ratio":0.008,"combined_severity":0.0037,"hygiene_severity":0.003,"hygiene_value_pct":99.7,"year":2018}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}